# TiffSaver: Only setLength of output file once