@dgault 2018-10-08
Attached show the before and after of the call stacks:
![profile_before](https://user-images.githubusercontent.com/1628865/46620876-e8801b80-cb1d-11e8-86b9-a4bc6fbcafa4.png)

After:
![profile_after](https://user-images.githubusercontent.com/1628865/46620888-ed44cf80-cb1d-11e8-91e5-1c9cb2ccc682.png)

Using bfconvert with tubhiswt-4D dataset:

Windows - Java 8
Without PR: 71.031s elapsed (4.6918607+74.8314ms per plane, 2125ms overhead)
With PR: 26.018s elapsed (4.816279+23.474419ms per plane, 1016ms overhead)

Windows - Java 10 (actually not impacted as much as expected)
Without PR: 78.485s
With PR: 23.783s

Tiling performance is not improved so that needs to be corrected.